refactor(errortracking): remove handled exceptions import cycle
