+++ ddtrace/errortracking/_handled_exceptions/callbacks.py
@@ -7,8 +7,8 @@ import traceback
 from ddtrace import config
 from ddtrace import tracer
 from ddtrace._trace.span import Span
-from ddtrace.errortracking._handled_exceptions.collector import HandledExceptionCollector
-from ddtrace.errortracking._handled_exceptions.collector import SpanEventData
+from ddtrace.errortracking._handled_exceptions.events import SpanEventData
+from ddtrace.errortracking._handled_exceptions.events import capture_exception_event
 from ddtrace.internal.utils.time import Time
 
 
@@ -72,4 +72,4 @@ def _default_errortracking_exc_callback(*args, span: Span, exc=None):
     generated = _generate_span_event(span, exc)
     if generated is not None:
         exc, span, span_event = generated
-        HandledExceptionCollector.capture_exception_event(span, exc, span_event)
+        capture_exception_event(span, exc, span_event)
